% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blav_tidiers.R
\name{glance.blavaan}
\alias{glance.blavaan}
\title{Glance at a blavaan object}
\usage{
\method{glance}{blavaan}(x, fit.indices = FALSE, ...)
}
\arguments{
\item{x}{A \code{blavaan} object}

\item{fit.indices}{Character vector of fit indices to compute from \code{fitMeasures}
and \code{blavFitIndices}. Use \code{"TRUE"} to also compute \code{blavFitIndices} measures.
Default is \code{"FALSE"}.}

\item{...}{Additional arguments (currently ignored).}
}
\value{
A single-row \code{data.frame}
  with columns:
  \item{npar}{Number of estimated parameters}
  \item{ngroups}{Number of groups}
  \item{ppp}{Posterior predictive p-value}
  \item{bic}{Bayesian information criterion}
  \item{dic}{Deviance information criterion}
  \item{waic}{Widely applicable information criterion}
  \item{looic}{Leave-one-out information criterion}
  \item{margloglik}{Marginal log-likelihood}
  Additional EAP of fit measures from \code{blavFitIndices} are included if requested.
}
\description{
Glance at a blavaan object
}
\examples{
\dontrun{
library(blavaan)

HS.model <- 'visual =~ x1 + x2 + x3'
fit <- bcfa(HS.model, data = HolzingerSwineford1939, seed = 123)
glance(fit)
glance(fit, fit.indices = TRUE)  # includes BRMSEA
}

}
